Renaming to interface
